backend/requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
pylance
pydantic
deprecated
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-dotenv
python-multipart
python-jose[cryptography]
email_validator
httpx
anyio
sniffio
greenlet
typing_extensions
annotated-types
passlib[bcrypt]
pg8000